refactor: controller에서 module 제거 및 값 얻는 메서드 분리
